requirements.txt
Flask
Pillow
gunicorn
redis
rq
tqdm
